fix: Decouple HSV color model from slider position
HSV color model doesn't allow hue to be valid when R=G=B and saturation
when value =0. This caused drop of H and S to 0 anytime when V is 0 and drop
of H when S is 0.
To prevent this, new class HSVSliders were introduced as stateless container of
HSV sliders to hold user-set slider positions.